changed some variable types in texturmanager.c from int to size_t for better memeory safety

+++ include/texturemanager.h
@@ -8,16 +8,16 @@
 
 typedef struct{
   SDL_Texture** textures;
-  int texture_capacity;
-  int texture_count;
+  size_t texture_capacity;
+  size_t texture_count;
   int texture_width;
   int texture_height;
 }TextureManager;
 
-TextureManager* texturemanager_create(int num_textures, SDL_Renderer* renderer, int texture_width, int texture_height);
+TextureManager* texturemanager_create(size_t num_textures, SDL_Renderer* renderer, int texture_width, int texture_height);
 int texturemanager_add_texture(TextureManager* texture_manager, SDL_Renderer* renderer, const char* texture_path);
-void texturemanager_remove_texture(TextureManager* texture_manager, int texutre_id);
-SDL_Texture* texturemanager_get_texture(TextureManager* texture_manager, int texture_id);
+void texturemanager_remove_texture(TextureManager* texture_manager, size_t texutre_id);
+SDL_Texture* texturemanager_get_texture(TextureManager* texture_manager, size_t texture_id);
 void texturemanager_destroy(TextureManager* texture_manager);
 
 #endif
